pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "An example project using Django & TimescaleDB"
authors = ["Rowan Molony <rowanmolony@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
